implemented dragging and discarding cards.
Discarding cards will remove the card from the player's hand and add it to the discard pile. It destroys the visual of the card, which means that the data of who was the card's owner before it was discarded will be lost. This should be fine, as far as I know we don't care whose card it used to be once it has been discarded.